# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"XK-Note API","description":"**This documentation describes the XK-Note API.**","contact":{"name":"Otstar Lin","email":"syfxlin@gmail.com"},"version":"0.1"},"paths":{"\/user":{"get":{"summary":"\u83b7\u53d6\u5f53\u524d\u7528\u6237\u4fe1\u606f","description":"\u8fd4\u56de\u5f53\u524d\u7528\u6237\u4fe1\u606f","operationId":"App\\Http\\Controllers\\API\\APIController::user","responses":{"200":{"description":"\u6210\u529f\u8fd4\u56de\u7528\u6237\u4fe1\u606f","content":{"application\/json":{"schema":{"properties":{"message":{"description":"\u54cd\u5e94\u4fe1\u606f","type":"string","example":"true or errorMassage"},"id":{"description":"\u7528\u6237ID","type":"int","example":"1"},"name":{"description":"\u7528\u6237\u540d","type":"string","example":"username"},"email":{"description":"\u7528\u6237\u90ae\u7bb1","type":"string","example":"user@email.com"},"api_token":{"description":"\u7528\u6237\u7684API Token","type":"string","example":"string"}},"type":"object"}}}},"301":{"description":"\u91cd\u5b9a\u5411\u5230\u767b\u9646\u754c\u9762"}}}},"\/admin\/users":{"post":{"summary":"\u521b\u5efa\u4e00\u4e2a\u7528\u6237","description":"\u521b\u5efa\u4e00\u4e2a\u7528\u6237","operationId":"App\\Http\\Controllers\\API\\APIController::createUser","requestBody":{"description":"\u7528\u6237\u4fe1\u606f","required":false,"content":{"application\/json":{"schema":{"properties":{"name":{"description":"\u7528\u6237\u540d","type":"string","example":"username"},"email":{"description":"\u7528\u6237\u90ae\u7bb1","type":"string","example":"user@email.com"},"limit":{"description":"\u9650\u5236\u521b\u5efa\u7684\u8d44\u6e90\u603b\u5927\u5c0f","type":"string","example":"10M"},"password":{"description":"\u7528\u6237\u5bc6\u7801","type":"string","example":"123456@abc"}},"type":"object"}}}},"responses":{"200":{"description":"\u6210\u529f\u521b\u5efa\u7528\u6237","content":{"application\/json":{"schema":{"properties":{"message":{"description":"\u54cd\u5e94\u4fe1\u606f","type":"string","example":"true or errorMassage"},"id":{"description":"\u7528\u6237ID","type":"int","example":"1"},"name":{"description":"\u7528\u6237\u540d","type":"string","example":"username"},"email":{"description":"\u7528\u6237\u90ae\u7bb1","type":"string","example":"user@email.com"},"api_token":{"description":"\u7528\u6237\u7684API Token","type":"string","example":"string"}},"type":"object"}}}},"301":{"description":"\u91cd\u5b9a\u5411\u5230\u767b\u9646\u754c\u9762"}}}}},"components":{"securitySchemes":{"api_key_security":{"type":"apiKey","description":"AccessToken","name":"access_token","in":"header"}}}}